config improvement + better model and block path
