+++ easydata/utils/__init__.py
@@ -1,0 +1,3 @@
+from easydata.config.loader import load_config
+
+config = load_config()
